[tools] perf pmu: Unbreak perf record for arm/arm64 with events with explicit PMU
Message-id: <20180419185703.28105-238-jolsa@redhat.com>
Patchwork-id: 210185
O-Subject: [PATCH RHEL7.6 BZ1540862 237/245] perf pmu: Unbreak perf record for arm/arm64 with events with explicit PMU
Bugzilla: 1540862
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1540862
https://bugzilla.redhat.com/show_bug.cgi?id=1540862

upstream
========
commit 66ec11919a0f96e936bb731fdbc2851316077d26
Author: Mark Rutland <mark.rutland@arm.com>
Date: Fri Oct 6 19:38:22 2017 +0100

description
===========
Currently, perf record is broken on arm/arm64 systems when the PMU is
specified explicitly as part of the event, e.g.

$ ./perf record -e armv8_cortex_a53/cpu_cycles/u true

In such cases, perf record fails to open events unless
perf_event_paranoid is set to -1, even if the PMU in question supports
mode exclusion. Further, even when perf_event_paranoid is toggled, no
samples are recorded.

This is an unintended side effect of commit:

  e3ba76deef23064f ("perf tools: Force uncore events to system wide monitoring)

... which assumes that if a PMU has an associated cpu_map, it is an
uncore PMU, and forces events for such PMUs to be system-wide.

This is not true for arm/arm64 systems, which can have heterogeneous
CPUs. To account for this, multiple CPU PMUs are exposed, each with a
"cpus" field under sysfs, which the perf tool parses into a cpu_map. ARM
PMUs do not have a "cpumask" file, and only have a "cpus" file. For the
gory details as to why, see commit:

 7e3fcffe95544010 ("perf pmu: Support alternative sysfs cpumask")

Given all of this, we can instead identify uncore PMUs by explicitly
checking for a "cpumask" file, and restore arm/arm64 PMU support back to
a working state. This patch does so, adding a new perf_pmu::is_uncore
field, and splitting the existing cpumask parsing so that it can be
reused.

Tested-by Will Deacon <will.deacon@arm.com>
Fixes: e3ba76deef23064f ("perf tools: Force uncore events to system wide monitoring)

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>