[x86] intel_rdt: Initialize bitmask of shareable resource if CDP enabled
Message-id: <20180425134743.26487-9-jolsa@redhat.com>
Patchwork-id: 211373
O-Subject: [PATCH RHEL7.6 BZ1509785 08/15] x86/intel_rdt: Initialize bitmask of shareable resource if CDP enabled
Bugzilla: 1509785
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1509785
https://bugzilla.redhat.com/show_bug.cgi?id=1509785

upstream
========
commit 95953034fb24c16ad0047a98b16427e5935830c4
Author: Reinette Chatre <reinette.chatre@intel.com>
Date: Fri Oct 20 02:16:57 2017 -0700

description
===========
The platform informs via CPUID.(EAX=0x10, ECX=res#):EBX[31:0] (valid res#
are only 1 for L3 and 2 for L2) which unit of the allocation may be used by
other entities in the platform. This information is valid whether CDP (Code
and Data Prioritization) is enabled or not.

Ensure that the bitmask of shareable resource is initialized when CDP is
enabled.

Fixes: 0dd2d7494cd8 ("x86/intel_rdt: Show bitmask of shareable resource with other executing units"
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>