[scsi] bnx2fc: Fix check in SCSI completion handler for timed out request
Message-id: <1524764221-2497-3-git-send-email-cdupuis@redhat.com>
Patchwork-id: 211611
O-Subject: [PATCH RHEL 7.6 e-stor 2/2] scsi: bnx2fc: Fix check in SCSI completion handler for timed out request
Bugzilla: 1566553
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Chris Leech <cleech@redhat.com>

From: Chad Dupuis <chad.dupuis@cavium.com>

commit ecf7ff49945f5741fa1da112f994939f942031d3

When a request times out we set the io_req flag BNX2FC_FLAG_IO_COMPL so
that if a subsequent completion comes in on that task ID we will ignore
it.  The issue is that in the check for this flag there is a missing
return so we will continue to process a request which may have already
been returned to the ownership of the SCSI layer.  This can cause
unpredictable results.

Solution is to add in the missing return.

[mkp: typo plus title shortening]

Signed-off-by: Chad Dupuis <chad.dupuis@cavium.com>
Reviewed-by: Laurence Oberman <loberman@redhat.com>
Tested-by: Laurence Oberman <loberman@redhat.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>