[netdrv] ibmvnic: Fix failover case for non-redundant configuration
Message-id: <1524921339-8359-5-git-send-email-drosario@redhat.com>
Patchwork-id: 211847
O-Subject: [RHEL-7.6 BZ1519746 PATCH 4/8] ibmvnic: Fix failover case for non-redundant configuration
Bugzilla: 1519746
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1519746

DESCRIPTION:
------------
There is a failover case for a non-redundant pseries VNIC
configuration that was not being handled properly. The current
implementation assumes that the driver will always have a redandant
device to communicate with following a failover notification. There
are cases, however, when a non-redundant configuration can receive
a failover request. If that happens, the driver should wait until
it receives a signal that the device is ready for operation.

The driver is agnostic of its backing hardware configuration,
so this fix necessarily affects all device failover management.
The driver needs to wait until it receives a signal that the device
is ready for resetting. A flag is introduced to track this intermediary
state where the driver is waiting for an active device.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 5a18e1e0c193b2f6a8d4651f38aaabee58080647

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net.git/commit/?id=5a18e1e0c193b2f6a8d4651f38aaabee58080647
----------------
Desnes A. Nunes do Rosario

PROPOSED PATCH:

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>