[netdrv] ibmvnic: Do not reset CRQ for Mobility driver resets
Message-id: <1524921339-8359-6-git-send-email-drosario@redhat.com>
Patchwork-id: 211845
O-Subject: [RHEL-7.6 BZ1519746 PATCH 5/8] ibmvnic: Do not reset CRQ for Mobility driver resets
Bugzilla: 1519746
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1519746

DESCRIPTION:
------------
When resetting the ibmvnic driver after a partition migration occurs
there is no requirement to do a reset of the main CRQ. The current
driver code does the required re-enable of the main CRQ, then does
a reset of the main CRQ later.

What we should be doing for a driver reset after a migration is to
re-enable the main CRQ, release all the sub-CRQs, and then allocate
new sub-CRQs after capability negotiation.

This patch updates the handling of mobility resets to do the proper
work and not reset the main CRQ. To do this the initialization/reset
of the main CRQ had to be moved out of the ibmvnic_init routine
and in to the ibmvnic_probe and do_reset routines.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 30f796258c49baa313222456bcf5b0246da55ff1

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net.git/commit/?id=30f796258c49baa313222456bcf5b0246da55ff1
----------------
Desnes A. Nunes do Rosario

PROPOSED PATCH:

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>