[netdrv] ibmvnic: Clear pending interrupt after device reset
Message-id: <1524921339-8359-8-git-send-email-drosario@redhat.com>
Patchwork-id: 211843
O-Subject: [RHEL-7.6 BZ1519746 PATCH 7/8] ibmvnic: Clear pending interrupt after device reset
Bugzilla: 1519746
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1519746

DESCRIPTION:
------------
Due to a firmware bug, the hypervisor can send an interrupt to a
transmit or receive queue just prior to a partition migration, not
allowing the device enough time to handle it and send an EOI. When
the partition migrates, the interrupt is lost but an "EOI-pending"
flag for the interrupt line is still set in firmware. No further
interrupts will be sent until that flag is cleared, effectively
freezing that queue. To workaround this, the driver will disable the
hardware interrupt and send an H_EOI signal prior to re-enabling it.
This will flush the pending EOI and allow the driver to continue
operation.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit f23e0643cd0b53e68e283b6f26194d56c28a2eb1

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net.git/commit/?id=f23e0643cd0b53e68e283b6f26194d56c28a2eb1
----------------
Desnes A. Nunes do Rosario

PROPOSED PATCH:

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>