[x86] perf/x86: Enable free running PEBS for REGS_USER/INTR
Message-id: <20180503110248.1193-3-jolsa@redhat.com>
Patchwork-id: 212351
O-Subject: [PATCH RHEL7.6 BZ1570591 002/152] perf/x86: Enable free running PEBS for REGS_USER/INTR
Bugzilla: 1570591
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1570591
https://bugzilla.redhat.com/show_bug.cgi?id=1570591

upstream
========
commit a47ba4d77e1236d214e5116b5631bc4c2d6e6369
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Aug 31 14:46:30 2017 -0700

description
===========
Currently free running PEBS is disabled when user or interrupt
registers are requested. Most of the registers are actually
available in the PEBS record and can be supported.

So we just need to check for the supported registers and then
allow it: it is all except for the segment register.

For user registers this only works when the counter is limited
to ring 3 only, so this also needs to be checked.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>