[uapi] perf/core: Add PERF_AUX_FLAG_COLLISION to report colliding samples
Message-id: <20180503110248.1193-10-jolsa@redhat.com>
Patchwork-id: 212218
O-Subject: [PATCH RHEL7.6 BZ1570591 009/152] perf/core: Add PERF_AUX_FLAG_COLLISION to report colliding samples
Bugzilla: 1570591
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1570591
https://bugzilla.redhat.com/show_bug.cgi?id=1570591

upstream
========
commit 085b30625e39df67d7320f22269796276c6b0c11
Author: Will Deacon <will.deacon@arm.com>
Date: Fri Sep 23 14:05:12 2016 +0100

description
===========
The ARM SPE architecture permits an implementation to ignore a sample
if the sample is due to be taken whilst another sample is already being
produced. In this case, it is desirable to report the collision to
userspace, as they may want to lower the sample period.

This patch adds a PERF_AUX_FLAG_COLLISION flag, so that such events can
be relayed to userspace.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>