[tools] perf tools: Provide mutex wrappers for pthreads rwlocks
Message-id: <20180503110248.1193-56-jolsa@redhat.com>
Patchwork-id: 212227
O-Subject: [PATCH RHEL7.6 BZ1570591 055/152] perf tools: Provide mutex wrappers for pthreads rwlocks
Bugzilla: 1570591
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1570591
https://bugzilla.redhat.com/show_bug.cgi?id=1570591

upstream
========
commit 0a7c74eae307894c6c95316c382f118aef8481e8
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Apr 4 13:15:04 2017 -0300

description
===========
Andi reported a performance drop in single threaded perf tools such as
'perf script' due to the growing number of locks being put in place to
allow for multithreaded tools, so wrap the POSIX threads rwlock routines
with the names used for such kinds of locks in the Linux kernel and then
allow for tools to ask for those locks to be used or not.

I.e. a tool may have a multithreaded phase and then switch to single
threaded, like the upcoming patches for the synthesizing of
PERF_RECORD_{FORK,MMAP,etc} for pre-existing processes to then switch to
single threaded mode in 'perf top'.

The init routines will not be conditional, this way starting as single
threaded to then move to multi threaded mode should be possible.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>