[tools] perf callchain: Store srcline in callchain_cursor_node
Message-id: <20180503110248.1193-84-jolsa@redhat.com>
Patchwork-id: 212272
O-Subject: [PATCH RHEL7.6 BZ1570591 083/152] perf callchain: Store srcline in callchain_cursor_node
Bugzilla: 1570591
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1570591
https://bugzilla.redhat.com/show_bug.cgi?id=1570591

upstream
========
commit 40a342cda2cd9bc8f7bf81c5ce1a141584760757
Author: Milian Wolff <milian.wolff@kdab.com>
Date: Mon Oct 9 22:32:56 2017 +0200

description
===========
This is mostly a preparation to enable the creation of full callchain
nodes for inline frames. Such frames will reference the IP of the
non-inlined frame, but hold the symbol and srcline for an inlined
location. As such, we won't be able to query the srcline on-demand based
on the IP alone. Instead, we will leverage the functionality provided by
this patch here, and store the srcline for the inlined nodes in the new
srcline member of callchain_cursor_node.

Note that this patch on its own leaks the srcline, as there is no
free_callchain_cursor_node or similar. A future patch will add caching
of the srcline and handle deletion properly.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>