[tools] perf callchain: Refactor inline_list to operate on symbols
Message-id: <20180503110248.1193-85-jolsa@redhat.com>
Patchwork-id: 212273
O-Subject: [PATCH RHEL7.6 BZ1570591 084/152] perf callchain: Refactor inline_list to operate on symbols
Bugzilla: 1570591
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1570591
https://bugzilla.redhat.com/show_bug.cgi?id=1570591

upstream
========
commit fea0cf842c7aa08950063264ab1cfbce4ba38c1b
Author: Milian Wolff <milian.wolff@kdab.com>
Date: Mon Oct 9 22:32:57 2017 +0200

description
===========
This is a requirement to create real callchain entries for inlined
frames.

Since the list of inlines usually contains the target symbol too, i.e.
the location where the frames get inlined to, we alias that symbol and
reuse it as-is is. This ensures that other dependent functionality keeps
working, most notably annotation of the target frames.

For all other entries in the inline_list, a fake symbol is created.
These are marked by new 'inlined' member which is set to true. Only
those symbols are managed by the inline_list and get freed when the
inline_list is deleted from within inline_node__delete.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>