[tools] perf callchain: Refactor inline_list to store srcline string directly
Message-id: <20180503110248.1193-86-jolsa@redhat.com>
Patchwork-id: 212278
O-Subject: [PATCH RHEL7.6 BZ1570591 085/152] perf callchain: Refactor inline_list to store srcline string directly
Bugzilla: 1570591
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1570591
https://bugzilla.redhat.com/show_bug.cgi?id=1570591

upstream
========
commit 2be8832f3c51cf9e36a3e80ff57f4137505c2ba4
Author: Milian Wolff <milian.wolff@kdab.com>
Date: Mon Oct 9 22:32:58 2017 +0200

description
===========
This is a preparation for the creation of real callchain entries for
inlined frames. The rest of the perf code uses the srcline string. As
such, using that also for the srcline API allows us to simplify some of
the upcoming code. Most notably, it will allow us to cache the srcline
for a given inline node and reuse it for different callchain entries.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>