[tools] perf top: Fix window dimensions change handling
Message-id: <20180503110248.1193-130-jolsa@redhat.com>
Patchwork-id: 212324
O-Subject: [PATCH RHEL7.6 BZ1570591 129/152] perf top: Fix window dimensions change handling
Bugzilla: 1570591
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1570591
https://bugzilla.redhat.com/show_bug.cgi?id=1570591

upstream
========
commit 89d0aeab4252adc2a7ea693637dd21c588bfa2d1
Author: Jiri Olsa <jolsa@kernel.org>
Date: Tue Nov 14 10:23:39 2017 +0100

description
===========
The stdio perf top crashes when we change the terminal
window size. The reason is that we assumed we get the
perf_top pointer as a signal handler argument which is
not the case.

Changing the SIGWINCH handler logic to change global
resize variable, which is checked in the main thread
loop.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>