[char] ipmi: retain ipmi_si for ipmi platform_device.name
Message-id: <1525354601-38937-1-git-send-email-tcamuso@redhat.com>
Patchwork-id: 212376
O-Subject: [RHEL7.6 1567691 1/1] ipmi: retain ipmi_si for ipmi platform_device.name
Bugzilla: 1567691
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1567691
brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15965310

RHEL-only

In kernel v3.17, an override field in struct platform_device was added
by upstream commit 3d713e0 to allow the driver for a device to be
specified which will override standard OF, ACPI, ID table, and name
matching, along with other changes in the driver core to support this
functionality.

This field was used in upstream ipmi commit 9f88145f, which refactored
DMI probing into its own statically linked module. In the new ipmi_dmi
module, the platform_device.name was changed to dmi-ipmi-si, and the
ssif name was changed to dmi-ipmi-ssif.

The upstream commit 9f88145f was backported in RHEL commit 6a2c99a
without the driver_override field, so the new platform_device.names
could not bind to the ipmi_si kmod.

This caused a panic at boot time when the attempt was made to load the
ipmi_si kmod for whch no platform_device could be found, leading to the
null pointer access violation.

Modifying the platform core to accommodate this small ipmi_si change has
kABI implications as well as the potential for collateral impact.

The least invasive solution is to retain the ipmi_si name for the ipmi
platform device so that it can be bound to the ipmi_si driver.

This was tested on platforms that deliver their ipmi information with
DMI and with both DMI and ACPI.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>

CC: Prarit Bhargava <prarit@redhat.com>
CC: David Arcari <arcari@redhat.com>
CC: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>