[platform] x86: intel-hid: Power button suspend on Dell Latitude 7275
Message-id: <20180506070529.20016-10-swood@redhat.com>
Patchwork-id: 212580
O-Subject: [RHEL7.6 PATCH 09/18] platform/x86: intel-hid: Power button suspend on Dell Latitude 7275
Bugzilla: 1445055
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=1445055

On Dell Latitude 7275 the 5-button array is not exposed in the ACPI
tables, but notififies are still sent to the Intel HID device object
(device ID INT33D5) in response to power button actions.  They were
ignored as the intel-hid driver was not prepared to take care of them
until recently.

Power button wakeup from suspend-to-idle was added in:

commit 635173a17b03 ("intel-hid: Wake up Dell Latitude 7275 from suspend-to-idle")

However power button suspend doesn't work yet on this platform so it
would be good to add it also.

On the affected platform (for which priv->array is NULL), add a new
upfront check against the power button press notification (0xCE) to
notify_handler(), outside the wakeup mode this time, which allows
reporting the power button press event and triggers the suspend. Also
catch and ignore the corresponding power button release notification
(0xCF) to stop it from being reported as an "unknown event" in the logs.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=196115
Tested-by: Jrme de Bretagne <jerome.debretagne@gmail.com>
Signed-off-by: Jrme de Bretagne <jerome.debretagne@gmail.com>
Acked-By: Mario Limonciello <mario.limonciello@dell.com>
[dvhart: minor coding style and commit message format fix]
Signed-off-by: Darren Hart (VMware) <dvhart@infradead.org>

Signed-off-by: Darren Hart (VMware) <dvhart@infradead.org>
(cherry picked from commit 821b85366284542e00dd834062144c637e818ee0)
Signed-off-by: Scott Wood <swood@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>

Conflicts:
	drivers/platform/x86/intel-hid.c