[security] Make the hardened user-copy code depend on having a hardened allocator
Message-id: <1525862132-8355-8-git-send-email-crecklin@redhat.com>
Patchwork-id: 213458
O-Subject: [RHEL7.6 BZ1562140 PATCH v4 07/22] Make the hardened user-copy code depend on having a hardened allocator
Bugzilla: 1562140
RH-Acked-by: Joe Lawrence <joe.lawrence@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

commit 6040e57658eee6eb1315a26119101ca832d1f854
       Author: Linus Torvalds <torvalds@linux-foundation.org>
       Date:   Fri Aug 19 12:47:01 2016 -0700

    Make the hardened user-copy code depend on having a hardened allocator

The kernel test robot reported a usercopy failure in the new hardened
sanity checks, due to a page-crossing copy of the FPU state into the
task structure.

This happened because the kernel test robot was testing with SLOB, which
doesn't actually do the required book-keeping for slab allocations, and
as a result the hardening code didn't realize that the task struct
allocation was one single allocation - and the sanity checks fail.

Since SLOB doesn't even claim to support hardening (and you really
shouldn't use it), the straightforward solution is to just make the
usercopy hardening code depend on the allocator supporting it.

Reported-by: kernel test robot <xiaolong.ye@intel.com>
Cc: Kees Cook <keescook@chromium.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>