[sound] alsa: seq: Don't allow resizing pool in use
Message-id: <15258632375081594-38-patchset-jkysela@redhat.com>
Patchwork-id: 213509
O-Subject: [RHEL 7.6 PATCH 037/555] ALSA: seq: Don't allow resizing pool in use
Bugzilla: 1535427
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

This is a fix for a (sort of) fallout in the recent commit
d15d662e89fc ("ALSA: seq: Fix racy pool initializations") for
CVE-2018-1000004.
As the pool resize deletes the existing cells, it may lead to a race
when another thread is writing concurrently, eventually resulting a
UAF.

A simple workaround is not to allow the pool resizing when the pool is
in use.  It's an invalid behavior in anyway.

Fixes: d15d662e89fc ("ALSA: seq: Fix racy pool initializations")
Reported-by:  <long7573@126.com>
Reported-by: Nicolai Stange <nstange@suse.de>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit d85739367c6d56e475c281945c68fdb05ca74b4c)
Bugzilla: 1535427
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>