[sound] alsa: seq: Process queue tempo/ppq change in a shot
Message-id: <15258632375081594-44-patchset-jkysela@redhat.com>
Patchwork-id: 213513
O-Subject: [RHEL 7.6 PATCH 043/555] ALSA: seq: Process queue tempo/ppq change in a shot
Bugzilla: 1535427
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

The SNDRV_SEQ_IOCTL_SET_QUEUE_TEMPO ioctl sets the tempo and the ppq
in a single call, while the current implementation updates each value
one by one.  This is a bit racy, and also suboptimal from the
performance POV, as each call does re-acquire the lock and invokes
the update of ALSA timer resolution.

This patch reorganizes the code slightly so that we change both the
tempo and the ppq in a shot.  The skew value can be put into the same
lock, but this is rather a rarely used feature and completely
independent from the temp/ppq (it's evaluated only in the interrupt),
so it's left as it was.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 671ec859e5ee06ab0bf968e639a25576b18865ad)
Bugzilla: 1535427
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>