[sound] alsa: asoc: da7219: clkdev_drop usage depends on CONFIG_COMMON_CLK
Message-id: <15258632375081594-331-patchset-jkysela@redhat.com>
Patchwork-id: 213809
O-Subject: [RHEL 7.6 PATCH 330/555] ALSA: ASoC: da7219: clkdev_drop usage depends on CONFIG_COMMON_CLK
Bugzilla: 1535427
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Fixes: ASoC: da7219: Add common clock usage for providing DAI clks

clkdev_drop usage in the codec remove function should be dependent
on if CONFIG_COMMON_CLK is defined for the platform, otherwise it
can cause build failures for platforms that do not support this.
The clkdev_* functions are still defined for those platforms, in
headers and source but the functions are not linked in.

This patch resolves this issue, so clkdev_drop is only used if
CONFIG_COMMON_CLK is defined.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 0f9d0e355d7c3fa5756313765a77961074017845)
Bugzilla: 1535427
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>