[sound] alsa: usb-audio: Allow to override the longname string
Message-id: <15258632375081594-518-patchset-jkysela@redhat.com>
Patchwork-id: 213990
O-Subject: [RHEL 7.6 PATCH 517/555] ALSA: usb-audio: Allow to override the longname string
Bugzilla: 1535427
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Historically USB-audio driver sets the card's longname field with the
details of the device and the bus information.  It's good per se, but
not preferable when it's referred as the identifier for UCM profile.

This patch adds a quirk profile_name field to override the card's
longname string to a pre-defined one, so that one can create a unique
and consistent ID string for the specific USB device via a quirk table
to be used as a UCM profile name.

The patch does a slight code refactoring to split out the functions to
set shortname and longname fields as well.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 07eca5fc3ebad1d33bc12a2f09670c0edd8e6eb6)
Bugzilla: 1535427
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>