[scsi] qla2xxx: Move #include qla_nvme.h to fix compile errors on RHEL 7
Message-id: <20180509135614.21599-19-hmadhani@redhat.com>
Patchwork-id: 214137
O-Subject: [RHEL 7.6 e-stor PATCH 018/143] qla2xxx: Move #include qla_nvme.h to fix compile errors on RHEL 7
Bugzilla: 1547714
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Move the #include of qla_nvme.h to after the system includes so that
qla_nvme.h has access to the scatter/gather list routines.

This patch is RHEL 7 specific as there are some block level changes
to the NVMe core which are not present in RHEL 7 including adding the
defines for processing scatter/gather lists. It was agreed offline that
the best to handle this is by moving the inclusion of qla_nvme.h to
after the scatterlist.h define so this patch set only makes changes to
the qla2xxx driver code itself.

[ RH only patch ]

Signed-off-by: Himanshu Madhani <hmadhani@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>