[nvme] nvme_fc: fix abort race on teardown with lld reject
Message-id: <1525888340-18642-55-git-send-email-dmilburn@redhat.com>
Patchwork-id: 214265
O-Subject: [RHEL7.6 BZ 1515584 54/75] nvme_fc: fix abort race on teardown with lld reject
Bugzilla: 1515584
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: James Smart <jsmart2021@gmail.com>

Another abort race: An io request is started, becomes active,
and is attempted to be started with the lldd. At the same time
the controller is stopped/torndown and an itterator is run to
abort the ios. As the io is active, it is added to the outstanding
aborted io count.  However on the original io request thread, the
driver ends up rejecting the io due to the condition that induced
the controller teardown. The driver reject path didn't check whether
it was in the outstanding io count. This left the count outstanding
stopping controller teardown.

Correct by, in the driver reject case, setting the state to
inactive and checking whether it was in the outstanding io count.

Signed-off-by: James Smart <james.smart@broadcom.com>
Reviewed-by: Sagi Grimberg <sagi@grimberg.me>
Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit b12740d316fa89f3f6191b71f986cf3b9383d379)
Signed-off-by: David Milburn <dmilburn@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>