[nvme] nvmet_fc: prevent new io rqsts in possible isr completions
Message-id: <1525888340-18642-57-git-send-email-dmilburn@redhat.com>
Patchwork-id: 214256
O-Subject: [RHEL7.6 BZ 1515584 56/75] nvmet_fc: prevent new io rqsts in possible isr completions
Bugzilla: 1515584
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: James Smart <jsmart2021@gmail.com>

When a bio completion calls back into the transport for a
back-end io device, the request completion path can free
the transport io job structure allowing it to be reused for
other operations. The transport has a defer_rcv queue which
holds temporary cmd rcv ops while waitng for io job structures.
when the job frees, if there's a cmd waiting, it is picked up
and submitted for processing, which can call back out to the
bio path if it's a read.  Unfortunately, what is unknown is the
context of the original bio done call, and it may be in a state
(softirq) that is not compatible with submitting the new bio in
the same calling sequence. This is especially true when using
scsi back-end devices as scsi is in softirq when it makes the
done call.

Correct by scheduling the io to be started via workq rather
than calling the start new io path inline to the original bio
done path.

Signed-off-by: James Smart <james.smart@broadcom.com>
Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit 9d625f7792875e8119ac3f364f3fd71b8bfc1294)
Signed-off-by: David Milburn <dmilburn@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>