[watchdog] hpwdt: Remove legacy NMI sourcing
Message-id: <1526068460-17717-5-git-send-email-jszczype@redhat.com>
Patchwork-id: 214704
O-Subject: [RHEL7.6 PATCH V3 4/7] watchdog: hpwdt: Remove legacy NMI sourcing.
Bugzilla: 1555073
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jerry Hoemann <jerry.hoemann@hpe.com>

BZ: 1555073

This commit was a cherry-pick that need some hand-holding
because of a couple of mis-matches between upstream and what's
in RHEL7, relative to the lines of code being removed by this
commit.

Upstream commit is at:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v4.16-rc6&id=2b3d89b402b085b08498e896c65267a145bed486

The few differences were all in areas of code that were being
removed anyhow, so the resolution was to delete all the lines of code
that were being deleted upstream, and to accept as-is the few
lines that were added.

Original text of upstrea commit follows:

Gen8 and prior Proliant systems supported the "CRU" interface
to firmware.  This interfaces allows linux to "call back" into firmware
to source the cause of an NMI.  This feature isn't fully utilized
as the actual source of the NMI isn't printed, the driver only
indicates that the source couldn't be determined when the call
fails.

With the advent of Gen9, iCRU replaces the CRU. The call back
feature is no longer available in firmware.  To be compatible and
not attempt to call back into firmware on system not supporting CRU,
the SMBIOS table is consulted to determine if it is safe to
make the call back or not.

This results in about half of the driver code being devoted
to either making CRU calls or determing if it is safe to make
CRU calls.  As noted, the driver isn't really using the results of
the CRU calls.

Furthermore, as a consequence of the Spectre security issue, the
BIOS/EFI calls are being wrapped into Spectre-disabling section.
Removing the call back in hpwdt_pretimeout assists in this effort.

As the CRU sourcing of the NMI isn't required for handling the
NMI and there are security concerns with making the call back, remove
the legacy (pre Gen9) NMI sourcing and the DMI code to determine if
the system had the CRU interface.

Signed-off-by: Jerry Hoemann <jerry.hoemann@hpe.com>
Acked-by: Ingo Molnar <mingo@kernel.org>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Wim Van Sebroeck <wim@iguana.be>
commit 2b3d89b402b085b08498e896c65267a145bed486

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>