[watchdog] hpwdt: Claim NMIs generated by iLO5
Message-id: <1526068460-17717-7-git-send-email-jszczype@redhat.com>
Patchwork-id: 214700
O-Subject: [RHEL7.6 PATCH V3 6/7] watchdog: hpwdt: Claim NMIs generated by iLO5
Bugzilla: 1555073
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jerry Hoemann <jerry.hoemann@hpe.com>

BZ: 1555073

This is a modified version of a patch sent upstream to lkml
for review.   It fixes and issue with NMIs not being claimed
when they should be.

The only difference here is that instead of using !mynmi, I
use !hpwdt_my_nmi() as we still use hpwdt_my_nmi() in RHEL7.

The lkml email for the fix is at:
https://lkml.org/lkml/2018/5/3/796

The upstream fix is currently in the hwmon git tree at:
https://git.kernel.org/pub/scm/linux/kernel/git/groeck/linux-staging.git/commit/?h=watchdog-next&id=9780372d9adccf79bb481a361f6867837b8ee0b3

This patch goes with this Red Hat commit.  The two should be
considered as one overall patch, as it addresses an issue in
that patch.  Other patch is:
commit aeada60ecf25f5673696af2fc57a579a7140f349
watchdog: hpwdt: condition early return of NMI handler on iLO5

Original text from this patch:
On iLO5 going forward we want to return and not claim the NMI, if
the NMI was NOT gnerated by the iLO as a result of the watchdog
timing out or an explicit generate NMI.

The sense of the test in is inverted and prevents hpwdt_pretimeout
from claiming NMIs when it should.

Signed-off-by: Jerry Hoemann <jerry.hoemann@hpe.com>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
commit 9780372d9adccf79bb481a361f6867837b8ee0b3

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>