[watchdog] hpwdt: bump driver version
Message-id: <1526068460-17717-8-git-send-email-jszczype@redhat.com>
Patchwork-id: 214703
O-Subject: [RHEL7.6 PATCH V3 7/7] watchdog: hpwdt: bump driver version
Bugzilla: 1555073
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ: 1555073

This is a RHEL7.6-only commit.

Update the hpwdt driver version from 1.4.0 to 1.4.0-RH1 to reflect the
updating of the driver from the 1.4.0 version that was present in
RHEL7.4.

It is not being updated to 2.0.0 because this update does not update
the driver (which is a driver rewrite) to conform with upstream linux'
watchdog core.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>