[netdrv] qed: Add unaligned and packed packet processing
Message-id: <1526390476-12354-25-git-send-email-cdupuis@redhat.com>
Patchwork-id: 215131
O-Subject: [PATCH v2 RHEL 7.6 24/74] qed: Add unaligned and packed packet processing
Bugzilla: 1548140
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Michal Kalderon <Michal.Kalderon@cavium.com>

Bugzilla : 1548140

commit 469981b17a4f8ddac91837bd74ebc98578f2ddbf

The fpdu data structure is preallocated per connection.
Each connection stores the current status of the connection:
either nothing pending, or there is a partial fpdu that is waiting for
the rest of the fpdu (incomplete bytes != 0).
The same structure is also used for splitting a packet when there are
packed fpdus. The structure is initialized with all data required
for sending the fpdu back to the FW. A fpdu will always be spanned across
a maximum of 3 tx bds. One for the header, one for the partial fdpu
received and one for the remainder (unaligned) packet.
In case of packed fpdu's, two fragments are used, one for the header
and one for the data.
Corner cases are not handled in the patch for clarity, and will be added
as a separate patch.

Signed-off-by: Michal Kalderon <Michal.Kalderon@cavium.com>
Signed-off-by: Ariel Elior <Ariel.Elior@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 469981b17a4f8ddac91837bd74ebc98578f2ddbf)
Signed-off-by: Harish Patil <hpatil@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>