[crypto] cryptd: Add helpers to check whether a tfm is queued
Message-id: <E1fJKib-0002lW-DO@gondobar>
Patchwork-id: 215477
O-Subject: [RHEL7.6 PATCH 2/2] crypto: cryptd - Add helpers to check whether a tfm is queued
Bugzilla: 1579195
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

RHEL7.6 Bugzilla #1579195

commit 81760ea6a95ad4c41273a71052f61b9f087b5753

This patch adds helpers to check whether a given tfm is currently
queued.  This is meant to be used by ablk_helper and similar
entities to ensure that no reordering is introduced because of
requests queued in cryptd with respect to requests being processed
in softirq context.

The per-cpu queue length limit is also increased to 1000 in line
with network limits.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>