[fs] NFSv4.1: Fix up replays of interrupted requests
Message-id: <20180517173539.6028-13-steved@redhat.com>
Patchwork-id: 215487
O-Subject: [RHEL7.6 fs BZ1575768 PATCH 12/12] NFSv4.1: Fix up replays of interrupted requests
Bugzilla: 1575768
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

If the previous request on a slot was interrupted before it was
processed by the server, then our slot sequence number may be out of whack,
and so we try the next operation using the old sequence number.

The problem with this, is that not all servers check to see that the
client is replaying the same operations as previously when they decide
to go to the replay cache, and so instead of the expected error of
NFS4ERR_SEQ_FALSE_RETRY, we get a replay of the old reply, which could
(if the operations match up) be mistaken by the client for a new reply.

To fix this, we attempt to send a COMPOUND containing only the SEQUENCE op
in order to resync our slot sequence number.

Cc: Olga Kornievskaia <olga.kornievskaia@gmail.com>
[olga.kornievskaia@gmail.com: fix an Oops]
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Signed-off-by: Anna Schumaker <Anna.Schumaker@Netapp.com>

(cherry picked from commit 3be0f80b5fe9c16eca2d538f799b94ca8aa59433)
Bugzilla: 1575768
---

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>