[scsi] qedf: Honor priority from DCBX FCoE App tag
Message-id: <1526913544-30231-12-git-send-email-cdupuis@redhat.com>
Patchwork-id: 216569
O-Subject: [PATCH RHEL 7.6 e-stor 11/23] scsi: qedf: Honor priority from DCBX FCoE App tag
Bugzilla: 1566547
RH-Acked-by: Chris Leech <cleech@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>

From: Chad Dupuis <chad.dupuis@cavium.com>

commit 84b2ba6e427ce57408052bde987260c022aff96c
https://kernel.googlesource.com/pub/scm/linux/kernel/git/jejb/scsi.git

We currently hard code the priority in the 8021q tag to 3 for FCoE
traffic.  The vast majority of the time this is fine but if the priority
is something else besides 3, any VLAN ID comparison either in the
non-offload path or offload path will fail and cause dropped frames
where none are expected.

Change the behavior so that the driver default is 3 if we do not get any
DCBX convergence.

If DCBX does converge, then set the FIP/FCoE priority in the following
manner:

 1. If the qedf_default_prio modparam is set use that
 2. If the DCBX FCoE priority is not in range (0..7) use 3
 3. Use the DCBX FCoE priority we get in the driver's DCBX handler

Signed-off-by: Chad Dupuis <chad.dupuis@cavium.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>