[x86] spec_ctrl: Use separate PCP variables for IBRS entry and exit
Message-id: <1526995149-19569-3-git-send-email-longman@redhat.com>
Patchwork-id: 216778
O-Subject: [RHEL7.6 PATCH 02/21] x86/spec_ctrl: Use separate PCP variables for IBRS entry and exit
Bugzilla: 1566905
CVE: CVE-2018-3639
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1566905
CVE: CVE-2018-3639 ("Speculative Store Bypass")

This is a RHEL7 specific patch.

The entry code that uses one percpu variable "spec_ctrl_pcp" for writing
the SPEC_CTRL MSR on kernel entry and exit is somewhat convoluted and
is hard to maintain when additional feature bits show up in that MSR.

To simplify the entry code and ease maintenance, The spec_ctrl_pcp
variable is now extended into a structure with the following fields:

 1) enabled
    - function like the original spec_ctrl_pcp to control action
      on kernel entry and exit.
 2) entry
    - lower 32-bit of SPEC_CTRL MSR to be written at kernel entry.
 3) exit
    - lower 32-bit of SPEC_CTRL MSR to be written at kernel exit.
 4) hi32
    - upper 32-bit of SPEC_CTRL MSR to be written at kernel entry/exit.

Another change is that for ibrs_always, writing of the SPEC_CTRL MSR
is now only done on kernel entry. That should improve the performance
of ibrs_always mode a bit by not writing to the MSR at kernel exit.

As a result of that change, there is no more 1-1 correspondance between
the IBRS mode and the value of the enabled field. So a global ibrs_mode
variable is now re-introduced.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>