[x86] spec_ctrl: Mask off SPEC_CTRL MSR bits that are managed by kernel
Message-id: <1526995149-19569-19-git-send-email-longman@redhat.com>
Patchwork-id: 216794
O-Subject: [RHEL7.6 PATCH 18/21] x86/spec_ctrl: Mask off SPEC_CTRL MSR bits that are managed by kernel
Bugzilla: 1566905
CVE: CVE-2018-3639
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1566905
CVE: CVE-2018-3639 ("Speculative Store Bypass")

This is a RHEL7 specific patch.

Running on a VM guest, it is possible that both the IBRS and SSBD bits
in the SPEC_CTRL MSR might have been set initially if they were set in
the host. This will cause those bits to be set in the x86_spec_ctrl_base
making those bits always on no matter the actual boot time or run time
setting chosen by the users.

As a result, we need to detect that and reset the MSRs and
x86_spec_ctrl_base to make sure that the kernel has full control of
what setting to be used as chosen by the users.

This patch also make sure that if SSBD is globally enabled, it is
written into the SPEC_CTRL MSRs in spec_ctrl_cpu_init();

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>