[x86] locking/barriers: Add nospec_barrier()
Message-id: <e295efca09a8d2dc6954c25bfc7720fc82813047.1527106687.git.jpoimboe@redhat.com>
Patchwork-id: 217143
O-Subject: [RHEL7.6 PATCH 5/8] locking/barriers: Add nospec_barrier()
Bugzilla: 1572653
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Christoph von Recklinghausen <crecklin@redhat.com>

This patch is RHEL-only.

In the Spectre embargo period, the variant 1 speculation barrier was
called gmb(), which is what it was called in RHEL.  Upstream it was
later renamed to nospec_barrier().  To ease future backports, add a
nospec_barrier() alias for gmb().

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>