[netdrv] fm10k: introduce a message queue for MAC/VLAN messages
Message-id: <20180524202230.23249-19-nhorman@redhat.com>
Patchwork-id: 217368
O-Subject: [RHEL7.6 PATCH 18/24] fm10k: introduce a message queue for MAC/VLAN messages
Bugzilla: 1518116
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Neil Horman <nhorman@redhat.com>

Under some circumstances, when dealing with a large number of MAC
address or VLAN updates at once, the fm10k driver, particularly the VFs
can overload the mailbox with too many messages at once.

This results in a mailbox timeout, which causes the driver to initiate
a reset. During the reset, we re-send all the same messages that
originally caused the timeout. This results in a cycle of resets each
triggering a future reset.

To fix or avoid this, we introduce a workqueue item which monitors
a queue of MAC and VLAN requests. These requests are queued to the end
of the list, and we process as a FIFO periodically.

Initially we only handle requests for the netdev, but we do handle
unicast MAC addresses, multicast MAC addresses, and update VLAN
requests.

A future patch will add support to use this queue for handling MAC
update requests from the VF<->PF mailbox.

The MAC/VLAN work item will keep checking to make sure that each request
does not overflow the mailbox and cause a timeout. If it might, then the
work item will reschedule itself a short time later. This avoids any
reset cycle, since we never send the message if the mailbox is not
ready.

As an alternative, we tried increasing the mailbox message FIFO, but
this just delays the problem and results in needless memory waste on the
system. Our new message queue is dynamically allocated so only uses as
much memory as it needs. Additionally, it need not be contiguous like
the Tx and Rx FIFOs.

Note that this patch chose to only create a queue for MAC and VLAN
messages, since these are the only messages sent in a large enough
volume to cause the reset loop. Other messages are very unlikely to
overflow the mailbox Tx FIFO so easily.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Tested-by: Krishneil Singh <krishneil.k.singh@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit fc9173682dcf73cfe3324267424ef17e854bb444)

Note: Little bit of divergence from upstream here due to a previously
skipped patch

BZ: 1518116
Testing: Tested successfully by intel
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=16449847
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>