[char] random: Wake up writers when random pools are zapped
Message-id: <20180527090445.zzqi53deii5pamyp@gondor.apana.org.au>
Patchwork-id: 217502
O-Subject: [RHEL7.6 PATCH] random: Wake up writers when random pools are zapped
Bugzilla: 1458419
RH-Acked-by: Bruno Eduardo de Oliveira Meneguele <bmeneg@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHEL7.6 Bugzilla #1458419
Brew https://brewweb.devel.redhat.com/taskinfo?taskID=16478832

This patch has been accepted into the maintainer's tree but isn't
in mainline yet so the commit ID may change.

https://git.kernel.org/pub/scm/linux/kernel/git/tytso/random.git/commit/?id=702e45cbb06dd3d4abe4672cde8ece89f130efaf

commit 702e45cbb06dd3d4abe4672cde8ece89f130efaf

As it is when the pool is zapped with RNDCLEARPOOL writers are
not woken up and therefore the pool may remain in the empty state
indefinitely.

This patch wakes them up unless the write threshold is set to zero.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Herbert Xu <herbert.xu@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>