[netdrv] bnxt_en: Improve valid bit checking in firmware response message
Message-id: <a619b629aec17f226789f00eab2d7e2b5eb0d2e8.1527638030.git.jtoppins@redhat.com>
Patchwork-id: 218316
O-Subject: [PATCH RHEL7.6 32/51] bnxt_en: Improve valid bit checking in firmware response message.
Bugzilla: 1526120
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

When firmware sends a DMA response to the driver, the last byte of the
message will be set to 1 to indicate that the whole response is valid.
The driver waits for the message to be valid before reading the message.

The firmware spec allows these response messages to increase in
length by adding new fields to the end of these messages.  The
older spec's valid location may become a new field in a newer
spec.  To guarantee compatibility, the driver should zero the valid
byte before interpreting the entire message so that any new fields not
implemented by the older spec will be read as zero.

For messages that are forwarded to VFs, we need to set the length
and re-instate the valid bit so the VF will see the valid response.

Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 845adfe40c2a75e67ddae6639fc2b987338b7983)
Bugzilla: 1526120
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=16505579
Tested: testing results from partner to be posted to bug, in addition did basic netperf testing with no performance loss noticed.
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>