[scsi] mpt3sas: SGL to PRP Translation for I/Os to NVMe devices
Message-id: <20180601150036.29582-6-thenzl@redhat.com>
Patchwork-id: 219685
O-Subject: [RHEL7.6 e-stor 05/60] scsi: mpt3sas: SGL to PRP Translation for I/Os to NVMe devices
Bugzilla: 1513855
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

* Added support for translating the SGLs associated with incoming
commands either to IEE SGL or NVMe PRPs for NVMe devices.

* The hardware translation of IEEE SGL to NVMe PRPs has limitations
and if a command cannot be translated by hardware then it will go to
firmware and the firmware needs to translate it. This will have a
performance impact. To avoid that, the driver proactively checks
whether the translation will be done in hardware or not. If not, then
driver translates.

[mkp: clarified commit message]

Signed-off-by: Chaitra P B <chaitra.basappa@broadcom.com>
Signed-off-by: Suganath Prabu S <suganath-prabu.subramani@broadcom.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 016d5c35e27824f31c394009dd0f72f2c6b0dc85)
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>