[netdrv] igb: Allow filters to be added for the local MAC address
Message-id: <20180605144316.32054-22-vinschen@redhat.com>
Patchwork-id: 220730
O-Subject: [RHEL7.6 PATCH 21/26] igb: Allow filters to be added for the local MAC address
Bugzilla: 1518105
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Vinicius Costa Gomes <vinicius.gomes@intel.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1518105
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=16600132

Users expect that when adding a steering filter for the local MAC
address, that all the traffic directed to that address will go to some
queue.

Currently, it's not possible to configure entries in the "in use"
state, which is the normal state of the local MAC address entry (it is
the default), this patch allows to override the steering configuration
of "in use" entries, if the filter to be added match the address and
address type (source or destination) of an existing entry.

There is a bit of a special handling for entries referring to the
local MAC address, when they are removed, only the steering
configuration is reset.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 872f923c5b38226a0c6825487742098b1968d463)
Signed-off-by: Corinna Vinschen <vinschen@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>