[kernel] bpf: Use mount_nodev not mount_ns to mount the bpf filesystem
Message-id: <20180612134439.10359-25-jolsa@redhat.com>
Patchwork-id: 221287
O-Subject: [PATCH RHEL7.6 BZ1311586 24/71] bpf: Use mount_nodev not mount_ns to mount the bpf filesystem
Bugzilla: 1311586
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit e27f4a942a0ee4b84567a3c6cfa84f273e55cbb7
Author: Eric W. Biederman <ebiederm@xmission.com>
Date: Fri May 20 17:22:48 2016 -0500

description
===========
While reviewing the filesystems that set FS_USERNS_MOUNT I spotted the
bpf filesystem.  Looking at the code I saw a broken usage of mount_ns
with current->nsproxy->mnt_ns. As the code does not acquire a
reference to the mount namespace it can not possibly be correct to
store the mount namespace on the superblock as it does.

Replace mount_ns with mount_nodev so that each mount of the bpf
filesystem returns a distinct instance, and the code is not buggy.

In discussion with Hannes Frederic Sowa it was reported that the use
of mount_ns was an attempt to have one bpf instance per mount
namespace, in an attempt to keep resources that pin resources from
hiding.  That intent simply does not work, the vfs is not built to
allow that kind of behavior.  Which means that the bpf filesystem
really is buggy both semantically and in it's implemenation as it does
not nor can it implement the original intent.

This change is userspace visible, but my experience with similar
filesystems leads me to believe nothing will break with a model of each
mount of the bpf filesystem is distinct from all others.

Fixes: b2197755b263 ("bpf: add support for persistent maps/progs")
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>