[kernel] perf, bpf: allow bpf programs attach to tracepoints
Message-id: <20180612134439.10359-68-jolsa@redhat.com>
Patchwork-id: 221322
O-Subject: [PATCH RHEL7.6 BZ1311586 67/71] perf, bpf: allow bpf programs attach to tracepoints
Bugzilla: 1311586
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 98b5c2c65c2951772a8fc661f50d675e450e8bce
Author: Alexei Starovoitov <ast@fb.com>
Date: Wed Apr 6 18:43:25 2016 -0700

description
===========
introduce BPF_PROG_TYPE_TRACEPOINT program type and allow it to be attached
to the perf tracepoint handler, which will copy the arguments into
the per-cpu buffer and pass it to the bpf program as its first argument.
The layout of the fields can be discovered by doing
'cat /sys/kernel/debug/tracing/events/sched/sched_switch/format'
prior to the compilation of the program with exception that first 8 bytes
are reserved and not accessible to the program. This area is used to store
the pointer to 'struct pt_regs' which some of the bpf helpers will use:
+---------+
| 8 bytes | hidden 'struct pt_regs *' (inaccessible to bpf program)
+---------+
| N bytes | static tracepoint fields defined in tracepoint/format (bpf readonly)
+---------+
| dynamic | __dynamic_array bytes of tracepoint (inaccessible to bpf yet)
+---------+

Not that all of the fields are already dumped to user space via perf ring buffer
and broken application access it directly without consulting tracepoint/format.
Same rule applies here: static tracepoint fields should only be accessed
in a format defined in tracepoint/format. The order of fields and
field sizes are not an ABI.

KABI
====
Changing the code in perf_trace_##call to take the struct bpf_prog
pointer from proper place.

Small conflicts in perf_trace_##call, dues to current version
of the function.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>