[crypto] af_alg: Forbid bind(2) when nokey child sockets are present
Message-id: <E1fT7LD-0007xF-3Z@gondobar>
Patchwork-id: 221418
O-Subject: [RHEL7.6 PATCH 2/2] crypto: af_alg - Forbid bind(2) when nokey child sockets are present
Bugzilla: 1555199
RH-Acked-by: Bruno Eduardo de Oliveira Meneguele <bmeneg@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

RHEL7.6 Bugzilla #1555199

commit a6a48c565f6f112c6983e2a02b1602189ed6e26e

This patch forbids the calling of bind(2) when there are child
sockets created by accept(2) in existence, even if they are created
on the nokey path.

This is needed as those child sockets have references to the tfm
object which bind(2) will destroy.

Cc: stable@vger.kernel.org
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>