[kernel] bpf: add BPF_PROG_ATTACH and BPF_PROG_DETACH commands
Message-id: <20180613214317.28803-5-jolsa@redhat.com>
Patchwork-id: 221475
O-Subject: [PATCH RHEL7.6 BZ1311586 004/386] bpf: add BPF_PROG_ATTACH and BPF_PROG_DETACH commands
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit f4324551489e8781d838f941b7aee4208e52e8bf
Author: Daniel Mack <daniel@zonque.org>
Date: Wed Nov 23 16:52:27 2016 +0100

description
===========
Extend the bpf(2) syscall by two new commands, BPF_PROG_ATTACH and
BPF_PROG_DETACH which allow attaching and detaching eBPF programs
to a target.

On the API level, the target could be anything that has an fd in
userspace, hence the name of the field in union bpf_attr is called
'target_fd'.

When called with BPF_ATTACH_TYPE_CGROUP_INET_{E,IN}GRESS, the target is
expected to be a valid file descriptor of a cgroup v2 directory which
has the bpf controller enabled. These are the only use-cases
implemented by this patch at this point, but more can be added.

If a program of the given type already exists in the given cgroup,
the program is swapped automically, so userspace does not have to drop
an existing program first before installing a new one, which would
otherwise leave a gap in which no program is attached.

For more information on the propagation logic to subcgroups, please
refer to the bpf cgroup controller implementation.

The API is guarded by CAP_NET_ADMIN.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>