[kernel] bpf: Mark __bpf_prog_run() stack frame as non-standard
Message-id: <20180613214317.28803-12-jolsa@redhat.com>
Patchwork-id: 221542
O-Subject: [PATCH RHEL7.6 BZ1311586 011/386] bpf: Mark __bpf_prog_run() stack frame as non-standard
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 39853cc0cdcf1b11f00f7f81e2f515a4d68ed209
Author: Josh Poimboeuf <jpoimboe@redhat.com>
Date: Sun Feb 28 22:22:37 2016 -0600

description
===========
objtool reports the following false positive warnings:

  kernel/bpf/core.o: warning: objtool: __bpf_prog_run()+0x5c: sibling call from callable instruction with changed frame pointer
  kernel/bpf/core.o: warning: objtool: __bpf_prog_run()+0x60: function has unreachable instruction
  kernel/bpf/core.o: warning: objtool: __bpf_prog_run()+0x64: function has unreachable instruction
  [...]

It's confused by the following dynamic jump instruction in
__bpf_prog_run()::

  jmp     *(%r12,%rax,8)

which corresponds to the following line in the C code:

  goto *jumptable[insn->code];

There's no way for objtool to deterministically find all possible
branch targets for a dynamic jump, so it can't verify this code.

In this case the jumps all stay within the function, and there's nothing
unusual going on related to the stack, so we can whitelist the function.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>